Create basic html, css and javacript
I've created the basic structure that make the
main function of the webpage work. Now you can
change the color of the background with just one
click on the button! I've created the basic html,
linked the css of the page and the javascript that
can make it all work.